+++ frontend/src/helper/compHelper.jsx
@@ -8,6 +8,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { addItem, delItem } from "../redux/product/productSlice";
 import Cart from "../views/Cart/Cart";
 import AuthService from "../services/AuthService";
+import { checkoutPost } from "./helper";
 export const LoadingProducts = () => {
   return (
     <>
@@ -206,6 +207,36 @@ export const ShowProducts = ({ results }) => {
         </>
       );
 };
+
+
+
+export async function checkoutFunction() {
+  const { value } = useSelector(
+    (state) => state.products
+  );
+  const userID = AuthService.getUserId();  
+  console.log(value);
+  console.log(value)
+  try {
+    const products = [];
+    value.forEach((res) => {
+      products.push({ productId : res._id , quantity : res.qty });
+    });
+    console.log(products);
+    const userData = {
+      "userID": userID,
+      "products": products,
+    };
+    console.log(userData);
+    const resp = await checkoutPost({ userData }); // Assuming checkoutPost is an async function
+    console.log(resp);
+    return resp;
+  } catch (e) {
+    console.error("Error in placing order:", e);
+    throw new Error(e);
+  }
+}
+
 
 export const ShowProduct = ({ product }) => {
   const dispatch = useDispatch();
